+++ alacritty/src/display/window.rs
@@ -60,7 +60,7 @@ use crate::gl;
 #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
 static WINDOW_ICON: &[u8] = include_bytes!("../../alacritty.png");
 
-/// This should match the definition of IDI_ICON from `windows.rc`.
+/// This should match the definition of IDI_ICON from `alacritty-icon.rc`.
 #[cfg(windows)]
 const IDI_ICON: WORD = 0x101;
 
@@ -233,6 +233,9 @@ impl Window {
         let current_mouse_cursor = CursorIcon::Text;
         windowed_context.window().set_cursor_icon(current_mouse_cursor);
 
+        // Enable IME.
+        windowed_context.window().set_ime_allowed(true);
+
         // Set OpenGL symbol loader. This call MUST be after window.make_current on windows.
         gl::load_with(|symbol| windowed_context.get_proc_address(symbol) as *const _);
 
@@ -331,6 +334,7 @@ impl Window {
 
         let builder = WindowBuilder::new()
             .with_title(&identity.title)
+            .with_name(&identity.class.instance, &identity.class.general)
             .with_visible(false)
             .with_transparent(true)
             .with_decorations(window_config.decorations != Decorations::None)
@@ -339,13 +343,6 @@ impl Window {
 
         #[cfg(feature = "x11")]
         let builder = builder.with_window_icon(icon.ok());
-
-        #[cfg(feature = "wayland")]
-        let builder = builder.with_app_id(identity.class.instance.to_owned());
-
-        #[cfg(feature = "x11")]
-        let builder = builder
-            .with_class(identity.class.instance.to_owned(), identity.class.general.to_owned());
 
         #[cfg(feature = "x11")]
         let builder = match &window_config.gtk_theme_variant {
@@ -461,6 +458,11 @@ impl Window {
         let nspot_y = f64::from(size.padding_y() + (point.line.0 + 1) as f32 * size.cell_height());
 
         self.window().set_ime_position(PhysicalPosition::new(nspot_x, nspot_y));
+    }
+
+    /// Enable or disable IME input for that window.
+    pub fn set_ime_allowed(&self, allowed: bool) {
+        self.window().set_ime_allowed(allowed);
     }
 
     pub fn swap_buffers(&self) {